Tue, Oct 3, 2023, 9:01 PM +07:00

+++ reverse/reverse.c
@@ -59,7 +59,7 @@ int main(int argc, char *argv[])
 
     // Write reversed audio to file
     // TODO #8
-    uint16_t *data = malloc(sizeof(block_size));
+    uint8_t *data = malloc(sizeof(block_size));
     while(fread(data, 1, block_size, input) == block_size)
     {
         for (int i = 0; i < block_size; i++)
